feat: 多 LLM 路由框架 + PDF 上傳解析 + 前後端整合修復
- 新增 LLMService 多供應商路由（Claude/OpenAI/Gemini 自動切換）
- ClaudeService 改為 LLMService backward-compatible wrapper
- AI 教練新增相關性判斷（LLM 語意檢查 + 不相關拒絕回答）
- PDF 上傳改為 multipart file + 背景非同步處理
- PDF 解析優先用本地 pymupdf，Claude API 作為 fallback
- Voyage AI embedding 失敗時靜默跳過（chunks 仍存入）
- 修復 config.py load_dotenv override 問題
- 修復 /onboarding/summary 缺少 subject id
- 修復 /resources API 欄位名與 ORM 不匹配
- 修復 ai_generation_service self._claude → self._llm
- 修復知識庫頁面前後端欄位映射
- 修復測驗設定頁面 documentService.list() 映射
- 修復考試 API 支援 document_ids
- 修復考試 workspace resume API 回傳 questions
- 修復雷達圖無資料時顯示假圖形
- 修復 Firebase 無 apiKey 時 console error
- TiTi Logo SVG + DBML 同步
- 新增 python-multipart、pymupdf 依賴

+++ backend/app/schemas/resource.py
@@ -7,7 +7,7 @@ from pydantic import BaseModel
 class UploadResourceRequest(BaseModel):
     filename: Optional[str] = None
     subject_id: Optional[str] = None
-    file_size_mb: Optional[int] = None
+    file_size_mb: Optional[float] = None
     type: Optional[str] = None
